Co-opt AUDIO packet to create channel status flag register on MOS
This change expands the use of the audio packet to set the nth bit in a global byte (channel_status) when audio is queued up, and then reset that byte when the audio is done playing. This means that "getsysvar_audioSuccess" now becomes a status register for all audio channels, not just the most recently played and you can query this on the ez80 side rather than guessing when a channel is clear based on delays.