Correct behaviour of VDU 28 (set text viewport)
Whilst porting MicroEMACS to Agon MOS I have been using VDU 28 to
simulate terminal action "clear to end of line" and "clear to end
of page" by creating an appropriately sized text viewport then
sending VDU 12 to clear it. I noticed it was not working very well
and performed some testing against the Linux version of BBC BASIC.
I observed  VDU 28 was not operating in the same way.  I have
updated the code to match the behaviour I was observing. In
particular, the ability to set a viewpoint a single line high that
ends at the end of the screen.

I have also made the code tolerant to speciying "right" and
"bottom" values that are off the screen - these are capped to the
maximum valid values. This also matches the behaviour I observed
with the Linux version of BBC BASIC.