improved sample handling
adds EnhancedSamplesGenerator which is used in place of the fab-gl SamplesGenerator

this accepts a audio_sample object for its sample data.  the setFrequency method is used to allow us to indicate to the generator that the index should be reset, as we can’t override the enable call

this approach will allow us to add looping information to the audio_sample object, and to use that in the EnhancedSamplesGenerator

deleting a sample removes it from any channels it is active on.