fix compiler warnings
fixes all compiler warnings by either removing unnecessary checks, fixing types to be more appropriate, and fixing a few minor bugs that these warnings were telling us about

one warning remains, which is to do with the RTC day of year value in our RTC VDP info packet.  this warning reflects a bug as we’re trying to send a day of year (range 0-365) in a single byte (range 0-255) which doesn’t fit.