fix mono bitmaps
create monochrome bitmaps using graphics foreground colour

also fixes logic for working out width in bytes of monochrome bitmaps.  bitmaps are assumed to have a complete number of bytes per row, and start at the top-most bit.  the width given is a pixel width, and does not have to be a multiple of 8.  the final byte in a row will have pixels ignored if the width is not a multiple of 8.