preliminary looping
support added for defining looping on a sample

sample playback however needs reworking.  the current system has the AudioSample structure keeping track of where playback is inside a sample, however that wouldn’t allow the same sample to be played simultaneously on more than one channel

control of sample playback instead needs to move into the EnhancedSamplesGenerator.  much of what’s in this commit needs to be moved